Moving/renaming FFmpeg log level
